+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+* Pom dependencies scopes update fixed:
+    - use both group and artifact for matching
+    - avoid marking as provided or optional transitive dependencies which also directly declared in compile
+
 ### 1.0.1 (2015-12-04)
 
 * Pom closure value now properly merged into existing pom (avoid duplicate tags)